// frontend/src/api/models/createPersonalAccessTokenResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Splice API
 * Financial account management API
 * OpenAPI spec version: 1.0
 */

export interface CreatePersonalAccessTokenResponse {
  id: string
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string
  /** @minLength 1 */
  token: string
  /** @minLength 1 */
  tokenPreview: string
  /** @nullable */
  expiresAt: string | null
  createdAt: string
}