// frontend/src/api/models/personalAccessToken.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Splice API
 * Financial account management API
 * OpenAPI spec version: 1.0
 */

export interface PersonalAccessToken {
  id: string
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string
  /** @minLength 1 */
  tokenPreview: string
  /** @nullable */
  lastUsedAt: string | null
  /** @nullable */
  expiresAt: string | null
  /** @nullable */
  revokedAt: string | null
  createdAt: string
}