// frontend/src/api/models/recurringManualTransactionSchedule.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Splice API
 * Financial account management API
 * OpenAPI spec version: 1.0
 */
import type { MoneyWithSign } from './moneyWithSign'
import type { RecurringManualTransactionScheduleCategory } from './recurringManualTransactionScheduleCategory'
import type { RecurringManualTransactionFrequency } from './recurringManualTransactionFrequency'

export interface RecurringManualTransactionSchedule {
  id: string
  accountId: string
  /** @nullable */
  accountName?: string | null
  amount: MoneyWithSign
  merchantName: string
  categoryId: string
  category?: RecurringManualTransactionScheduleCategory
  frequency: RecurringManualTransactionFrequency
  /**
   * @minimum 1
   * @maximum 31
   */
  dayOfMonth: number
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  startDate: string
  /**
   * @nullable
   * @pattern ^\d{4}-\d{2}-\d{2}$
   */
  endDate: string | null
  /**
   * @nullable
   * @pattern ^\d{4}-\d{2}-\d{2}$
   */
  nextOccurrenceDate: string | null
  /**
   * @nullable
   * @pattern ^\d{4}-\d{2}-\d{2}$
   */
  lastGeneratedOccurrenceDate: string | null
  /** @nullable */
  pausedAt: string | null
  /** @nullable */
  archivedAt: string | null
  createdAt: string
  updatedAt: string
  userId: string
}